pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nanowm"
version = "0.1.0"
description = "simple world models"
authors = [{ name="punwai" }]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0", 
    "pillow>=9.0.0",
    "datasets>=2.0.0",
    "einops>=0.6.0"
]

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["*"]  # package names should match these glob patterns